requirements.txt
Django==4.2.21
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
bcrypt==4.1.2
django-filter==24.1 